Move raid models into seperate files #2
- Reset to original and did not auto format files.
