+++ Raidbot/Raidbot.csproj
@@ -17,7 +17,7 @@
 
   <ItemGroup>
     <None Update="debugtoken.txt">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
     <None Update="RoleMessages.xml">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
